More default argument options
Previously, default arguments had to be specified with "uint8 age = 1;"
Now, you no longer have to specify a default argument and can just write
"uint8 age;". This will remove the default parameter from the
constructor. This may be helpful as it will force the user to set fields
before sending.  The call to the constructor in ABSPacket::Create() will
pass common default values into the constructor.  For example, the dog
field would be set to 0, and strings will be initialized as empty.  You
can change the values that are passed into the constructor in the
Create() method by writing "uint8 dog ~ 120;" This would set the dog
field as 120 instead of 0.